Move RemoteCryptoMarketPricesReceiver to a file
